+++ core/src/main/groovy/com/virtusa/gto/nyql/Where.groovy
@@ -4,6 +4,7 @@ import com.virtusa.gto.nyql.exceptions.NyScriptNotFoundException
 import com.virtusa.gto.nyql.exceptions.NySyntaxException
 import com.virtusa.gto.nyql.model.QScript
 import com.virtusa.gto.nyql.model.units.AParam
+import com.virtusa.gto.nyql.model.units.ParamBinary
 import com.virtusa.gto.nyql.model.units.ParamDate
 import com.virtusa.gto.nyql.model.units.ParamList
 import com.virtusa.gto.nyql.model.units.ParamTimestamp
@@ -81,6 +82,11 @@ class Where implements DataTypeTraits, FunctionTraits, ScriptTraits {
     @Override
     AParam PARAM_TIMESTAMP(String name, String format) {
         _ctx.addParam(new ParamTimestamp(__name: name, __tsFormat: format))
+    }
+
+    @Override
+    AParam PARAM_BINARY(String name) {
+        _ctx.addParam(new ParamBinary(__name: name))
     }
 
     @CompileStatic