+++ core/src/main/groovy/com/virtusa/gto/nyql/model/units/ParamBinary.groovy
@@ -1,0 +1,15 @@
+package com.virtusa.gto.nyql.model.units
+
+import groovy.transform.CompileStatic
+
+/**
+ * @author Isuru Weerarathna
+ */
+@CompileStatic
+class ParamBinary extends AParam {
+
+    @Override
+    boolean __shouldValueConvert() {
+        true
+    }
+}